requirements.txt
google.generativeai
python-dotenv
torch
tensorflow
transformers
